626763: fix ordinal initialization
getFields makes no guarantee about the order of the returned fields.
We need to specify an order to the statically initialized dynamic enumerations despite of that.
We will explicitly specify these values, as they need to have standardized values (which happen
to be the same as in the ODS standard in many cases)
This requires small modifications to the constructors of predefined enumerations, as well as to the enumeration and enumerationvalue classes.

Change-Id: I3e534645d45a6566913d4792c43195ea77bb67be
Signed-off-by: Florian Schmitt <florian.schmitt@atos.net>